ref(scheduler): move public method to top of k8s class and private methods grouped by resource type
